+++ backend/windmill-common/src/lib.rs
@@ -35,6 +35,7 @@ pub mod apps;
 pub mod audit;
 pub mod assets;
 pub mod auth;
+pub mod runtime_assets;
 #[cfg(feature = "benchmark")]
 pub mod bench;
 pub mod cache;
@@ -59,6 +60,7 @@ pub mod job_metrics;
 #[cfg(all(feature = "parquet", feature = "private"))]
 pub mod job_s3_helpers_ee;
 pub mod min_version;
+pub mod notify_events;
 #[cfg(feature = "parquet")]
 pub mod job_s3_helpers_oss;
 pub mod workspace_dependencies;